# Query Distinct : Add Performance Benchmarks and Improvements