--- a/Microsoft.Azure.Cosmos/src/Number64.cs
+++ b/Microsoft.Azure.Cosmos/src/Number64.cs
@@ -23,11 +23,6 @@ namespace Microsoft.Azure.Cosmos
     readonly struct Number64 : IComparable<Number64>, IEquatable<Number64>
     {
         /// <summary>
-        /// Size of the Number64 when compared as a DoubleEx
-        /// </summary>
-        public const int SizeOf = sizeof(double) + sizeof(ushort);
-
-        /// <summary>
         /// Maximum Number64.
         /// </summary>
         public static readonly Number64 MaxValue = new Number64(double.MaxValue);
@@ -119,20 +114,6 @@ namespace Microsoft.Azure.Cosmos
             }
 
             return toString;
-        }
-
-        public void CopyTo(Span<byte> buffer)
-        {
-            if (buffer.Length < Number64.SizeOf)
-            {
-                throw new ArgumentException($"{nameof(buffer)} is too short.");
-            }
-
-            DoubleEx doubleEx = Number64.ToDoubleEx(this);
-            Span<double> doubleBuffer = MemoryMarshal.Cast<byte, double>(buffer);
-            doubleBuffer[0] = doubleEx.DoubleValue;
-            Span<ushort> ushortBuffer = MemoryMarshal.Cast<byte, ushort>(buffer.Slice(sizeof(double)));
-            ushortBuffer[0] = doubleEx.ExtraBits;
         }
 
         #region Static Operators
@@ -253,7 +234,7 @@ namespace Microsoft.Azure.Cosmos
             return value;
         }
 
-        private static DoubleEx ToDoubleEx(Number64 number64)
+        public static DoubleEx ToDoubleEx(Number64 number64)
         {
             DoubleEx doubleEx;
             if (number64.IsDouble)
@@ -363,7 +344,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Represents an extended double number with 62-bit mantissa which is capable of representing a 64-bit integer with no precision loss
         /// </summary>
-        private readonly struct DoubleEx : IEquatable<DoubleEx>, IComparable<DoubleEx>
+        public readonly struct DoubleEx : IEquatable<DoubleEx>, IComparable<DoubleEx>
         {
             private DoubleEx(double doubleValue, ushort extraBits)
             {