--- a/Microsoft.Azure.Cosmos/src/Query/Core/ExecutionComponent/Distinct/DistinctHash.cs
+++ b/Microsoft.Azure.Cosmos/src/Query/Core/ExecutionComponent/Distinct/DistinctHash.cs
@@ -39,29 +39,26 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
         {
             public static readonly CosmosElementHasher Singleton = new CosmosElementHasher();
 
-            private static readonly UInt128 NullHashSeed = UInt128.Create(0x1380f68bb3b0cfe4, 0x156c918bf564ee48);
+            private static class HashSeeds
+            {
+                public static readonly UInt128 Null = UInt128.Create(0x1380f68bb3b0cfe4, 0x156c918bf564ee48);
+                public static readonly UInt128 False = UInt128.Create(0xc1be517fe893b40c, 0xe9fc8a4c531cd0dd);
+                public static readonly UInt128 True = UInt128.Create(0xf86d4abf9a412e74, 0x788488365c8a985d);
+                public static readonly UInt128 String = UInt128.Create(0x61f53f0a44204cfb, 0x09481be8ef4b56dd);
+                public static readonly UInt128 Array = UInt128.Create(0xfa573b014c4dc18e, 0xa014512c858eb115);
+                public static readonly UInt128 Object = UInt128.Create(0x77b285ac511aef30, 0x3dcf187245822449);
+                public static readonly UInt128 ArrayIndex = UInt128.Create(0xfe057204216db999, 0x5b1cc3178bd9c593);
+                public static readonly UInt128 PropertyName = UInt128.Create(0xc915dde058492a8a, 0x7c8be2eba72e4634);
+                public static readonly UInt128 Binary = UInt128.Create(0x54841d59fe1ea46c, 0xd4edb0ba5c59766b);
+                public static readonly UInt128 Guid = UInt128.Create(0x53b5b8939b790f4b, 0x7cc5e09441fd6cb1);
+            }
 
-            private static readonly UInt128 FalseHashSeed = UInt128.Create(0xc1be517fe893b40c, 0xe9fc8a4c531cd0dd);
-
-            private static readonly UInt128 TrueHashSeed = UInt128.Create(0xf86d4abf9a412e74, 0x788488365c8a985d);
-
-            private static readonly UInt128 StringHashSeed = UInt128.Create(0x61f53f0a44204cfb, 0x09481be8ef4b56dd);
-
-            private static readonly UInt128 ArrayHashSeed = UInt128.Create(0xfa573b014c4dc18e, 0xa014512c858eb115);
-
-            private static readonly UInt128 ObjectHashSeed = UInt128.Create(0x77b285ac511aef30, 0x3dcf187245822449);
-
-            private static readonly UInt128 ArrayIndexHashSeed = UInt128.Create(0xfe057204216db999, 0x5b1cc3178bd9c593);
-
-            private static readonly UInt128 PropertyNameHashSeed = UInt128.Create(0xc915dde058492a8a, 0x7c8be2eba72e4634);
-
-            private static readonly UInt128 BinaryHashSeed = UInt128.Create(0x54841d59fe1ea46c, 0xd4edb0ba5c59766b);
-
-            private static readonly UInt128 GuidHashSeed = UInt128.Create(0x53b5b8939b790f4b, 0x7cc5e09441fd6cb1);
-
-            private static readonly UInt128 RootNullHash = MurmurHash3.Hash128(NullHashSeed, RootHashSeed);
-            private static readonly UInt128 RootFalseHash = MurmurHash3.Hash128(FalseHashSeed, RootHashSeed);
-            private static readonly UInt128 RootTrueHash = MurmurHash3.Hash128(TrueHashSeed, RootHashSeed);
+            private static class RootCache
+            {
+                public static readonly UInt128 Null = MurmurHash3.Hash128(HashSeeds.Null, RootHashSeed);
+                public static readonly UInt128 False = MurmurHash3.Hash128(HashSeeds.False, RootHashSeed);
+                public static readonly UInt128 True = MurmurHash3.Hash128(HashSeeds.True, RootHashSeed);
+            }
 
             private CosmosElementHasher()
             {
@@ -71,7 +68,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             public UInt128 Visit(CosmosArray cosmosArray, UInt128 seed)
             {
                 // Start the array with a distinct hash, so that empty array doesn't hash to another value.
-                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.ArrayHashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.Array, seed);
 
                 // Incorporate all the array items into the hash.
                 for (int index = 0; index < cosmosArray.Count; index++)
@@ -83,7 +80,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
                     // [true, false, true] and [true, true, false]
                     // due to the way the seed works.
                     // But we add the index just incase that property does not hold in the future.
-                    UInt128 arrayItemSeed = CosmosElementHasher.ArrayIndexHashSeed + index;
+                    UInt128 arrayItemSeed = CosmosElementHasher.HashSeeds.ArrayIndex + index;
                     hash = MurmurHash3.Hash128(hash, arrayItem.Accept(this, arrayItemSeed));
                 }
 
@@ -93,9 +90,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             public UInt128 Visit(CosmosBinary cosmosBinary, UInt128 seed)
             {
                 // Hash with binary seed to differntiate between empty binary and no binary.
-                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.BinaryHashSeed, seed);
-
-                // TODO: replace this with Span based hashing.
+                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.Binary, seed);
                 hash = MurmurHash3.Hash128(cosmosBinary.Value.Span, hash);
                 return hash;
             }
@@ -104,17 +99,17 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             {
                 if (seed == RootHashSeed)
                 {
-                    return cosmosBoolean.Value ? RootFalseHash : RootTrueHash;
+                    return cosmosBoolean.Value ? CosmosElementHasher.HashSeeds.True : CosmosElementHasher.HashSeeds.False;
                 }
 
                 return MurmurHash3.Hash128(
-                    cosmosBoolean.Value ? CosmosElementHasher.TrueHashSeed : CosmosElementHasher.FalseHashSeed,
+                    cosmosBoolean.Value ? CosmosElementHasher.HashSeeds.True : CosmosElementHasher.HashSeeds.False,
                     seed);
             }
 
             public UInt128 Visit(CosmosGuid cosmosGuid, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.GuidHashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.Guid, seed);
                 hash = MurmurHash3.Hash128(cosmosGuid.Value.ToByteArray(), hash);
                 return hash;
             }
@@ -123,10 +118,10 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             {
                 if (seed == RootHashSeed)
                 {
-                    return RootNullHash;
+                    return RootCache.Null;
                 }
 
-                return MurmurHash3.Hash128(CosmosElementHasher.NullHashSeed, seed);
+                return MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.Null, seed);
             }
 
             public UInt128 Visit(CosmosNumber cosmosNumber, UInt128 seed)
@@ -137,7 +132,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             public UInt128 Visit(CosmosObject cosmosObject, UInt128 seed)
             {
                 // Start the object with a distinct hash, so that empty object doesn't hash to another value.
-                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.ObjectHashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.Object, seed);
 
                 //// Intermediate hashes of all the properties, which we don't want to xor with the final hash
                 //// otherwise the following will collide:
@@ -171,7 +166,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
                 // This is consistent with equality comparison.
                 foreach (KeyValuePair<string, CosmosElement> kvp in cosmosObject)
                 {
-                    UInt128 nameHash = CosmosString.Create(kvp.Key).Accept(this, CosmosElementHasher.PropertyNameHashSeed);
+                    UInt128 nameHash = CosmosString.Create(kvp.Key).Accept(this, CosmosElementHasher.HashSeeds.PropertyName);
                     UInt128 propertyHash = kvp.Value.Accept(this, nameHash);
 
                     //// xor is symmetric meaning that a ^ b = b ^ a
@@ -194,7 +189,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosString cosmosString, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.StringHashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosElementHasher.HashSeeds.String, seed);
                 if (cosmosString.TryGetBufferedValue(out Utf8Memory bufferedUtf8Value))
                 {
                     hash = MurmurHash3.Hash128(bufferedUtf8Value.Span.Span, hash);
@@ -212,14 +207,17 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
         {
             public static readonly CosmosNumberHasher Singleton = new CosmosNumberHasher();
 
-            private static readonly UInt128 Number64HashSeed = UInt128.Create(0x2400e8b894ce9c2a, 0x790be1eabd7b9481);
-            private static readonly UInt128 Float32HashSeed = UInt128.Create(0x881c51c28fb61016, 0x1decd039cd24bd4b);
-            private static readonly UInt128 Float64HashSeed = UInt128.Create(0x62fb48cc659963a0, 0xe9e690779309c403);
-            private static readonly UInt128 Int8HashSeed = UInt128.Create(0x0007978411626daa, 0x89933677a85444b7);
-            private static readonly UInt128 Int16HashSeed = UInt128.Create(0xe7a19001d3211c09, 0x33e0ba9fb8bc7940);
-            private static readonly UInt128 Int32HashSeed = UInt128.Create(0x0320dc908e0d3e71, 0xf575de218f09ffa5);
-            private static readonly UInt128 Int64HashSeed = UInt128.Create(0xed93baf7fdc76638, 0x0d5733c37e079869);
-            private static readonly UInt128 UInt32HashSeed = UInt128.Create(0x78c441a2d2e9bb6e, 0xac88cb880ccda71d);
+            public static class HashSeeds
+            {
+                public static readonly UInt128 Number64 = UInt128.Create(0x2400e8b894ce9c2a, 0x790be1eabd7b9481);
+                public static readonly UInt128 Float32 = UInt128.Create(0x881c51c28fb61016, 0x1decd039cd24bd4b);
+                public static readonly UInt128 Float64 = UInt128.Create(0x62fb48cc659963a0, 0xe9e690779309c403);
+                public static readonly UInt128 Int8 = UInt128.Create(0x0007978411626daa, 0x89933677a85444b7);
+                public static readonly UInt128 Int16 = UInt128.Create(0xe7a19001d3211c09, 0x33e0ba9fb8bc7940);
+                public static readonly UInt128 Int32 = UInt128.Create(0x0320dc908e0d3e71, 0xf575de218f09ffa5);
+                public static readonly UInt128 Int64 = UInt128.Create(0xed93baf7fdc76638, 0x0d5733c37e079869);
+                public static readonly UInt128 UInt32 = UInt128.Create(0x78c441a2d2e9bb6e, 0xac88cb880ccda71d);
+            }
 
             private CosmosNumberHasher()
             {
@@ -228,7 +226,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosFloat32 cosmosFloat32, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Float32HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Float32, seed);
                 float value = cosmosFloat32.GetValue();
 
                 // Normalize 0.0f and -0.0f value
@@ -244,7 +242,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosFloat64 cosmosFloat64, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Float64HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Float64, seed);
                 double value = cosmosFloat64.GetValue();
 
                 // Normalize 0.0 and -0.0 value
@@ -260,7 +258,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosInt16 cosmosInt16, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Int16HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Int16, seed);
                 short value = cosmosInt16.GetValue();
                 hash = MurmurHash3.Hash128(value, hash);
                 return hash;
@@ -268,7 +266,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosInt32 cosmosInt32, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Int32HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Int32, seed);
                 int value = cosmosInt32.GetValue();
                 hash = MurmurHash3.Hash128(value, hash);
                 return hash;
@@ -276,7 +274,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosInt64 cosmosInt64, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Int64HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Int64, seed);
                 long value = cosmosInt64.GetValue();
                 hash = MurmurHash3.Hash128(value, hash);
                 return hash;
@@ -284,7 +282,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosInt8 cosmosInt8, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Int8HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Int8, seed);
                 sbyte value = cosmosInt8.GetValue();
                 hash = MurmurHash3.Hash128(value, hash);
                 return hash;
@@ -292,7 +290,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosNumber64 cosmosNumber64, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.Number64HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.Number64, seed);
                 Number64 value = cosmosNumber64.GetValue();
                 Span<byte> buffer = stackalloc byte[Number64.SizeOf];
                 value.CopyTo(buffer);
@@ -301,7 +299,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
 
             public UInt128 Visit(CosmosUInt32 cosmosUInt32, UInt128 seed)
             {
-                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.UInt32HashSeed, seed);
+                UInt128 hash = MurmurHash3.Hash128(CosmosNumberHasher.HashSeeds.UInt32, seed);
                 uint value = cosmosUInt32.GetValue();
                 hash = MurmurHash3.Hash128(value, hash);
                 return hash;