--- a/Microsoft.Azure.Cosmos/src/Query/Core/ExecutionComponent/Distinct/DistinctHash.cs
+++ b/Microsoft.Azure.Cosmos/src/Query/Core/ExecutionComponent/Distinct/DistinctHash.cs
@@ -104,7 +104,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Distinct
             {
                 if (seed == RootHashSeed)
                 {
-                    return cosmosBoolean.Value ? HashSeeds.True : HashSeeds.False;
+                    return cosmosBoolean.Value ? RootCache.True : RootCache.False;
                 }
 
                 return MurmurHash3.Hash128(