… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Query/DistinctHashBenchmarkResults.txt
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/BaselineTest/TestBaseline/DistinctHashBaselineTests.ElementsHash.xml
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/BaselineTest/TestBaseline/DistinctHashBaselineTests.ExtendedTypesHash.xml
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/BaselineTest/TestBaseline/DistinctHashBaselineTests.WrappedElementsHash.xml
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/testbaselines.cmd