# Query Distinct : Add Performance Benchmarks and Improvements

## Description

This PR adds a micro benchmark for taking the hash of different types of CosmosElements, which is a core operation for DISTINCT. Performance improvements were also added and the improvements are detailed here:

```
Initial:

|          Method |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|---------------- |----------:|----------:|----------:|----------:|------:|------:|-----------:|
|            Null |  2.580 ms | 0.0212 ms | 0.0188 ms |         - |     - |     - |          - |
|           False |  2.568 ms | 0.0173 ms | 0.0153 ms |         - |     - |     - |       20 B |
|            True |  2.563 ms | 0.0175 ms | 0.0155 ms |         - |     - |     - |        4 B |
|          Number |  7.216 ms | 0.0856 ms | 0.0801 ms |         - |     - |     - |          - |
|          String |  8.231 ms | 0.0762 ms | 0.0713 ms |         - |     - |     - |          - |
|      EmptyArray |  2.883 ms | 0.0164 ms | 0.0154 ms |         - |     - |     - |       13 B |
|  ArrayWithItems | 37.303 ms | 0.5150 ms | 0.4817 ms |         - |     - |     - |          - |
|     EmptyObject |  5.703 ms | 0.0239 ms | 0.0199 ms | 1140.6250 |     - |     - |  4800043 B |
| ObjectWithItems | 89.685 ms | 0.4908 ms | 0.4591 ms | 4833.3333 |     - |     - | 20801451 B |

Singleton Cache:

|          Method |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|---------------- |----------:|----------:|----------:|----------:|------:|------:|-----------:|
|            Null |  1.394 ms | 0.0274 ms | 0.0487 ms |         - |     - |     - |          - |
|           False |  1.391 ms | 0.0108 ms | 0.0101 ms |         - |     - |     - |        9 B |
|            True |  1.397 ms | 0.0109 ms | 0.0097 ms |         - |     - |     - |          - |
|          Number |  7.208 ms | 0.0456 ms | 0.0426 ms |         - |     - |     - |          - |
|          String |  8.274 ms | 0.1027 ms | 0.0858 ms |         - |     - |     - |          - |
|      EmptyArray |  2.883 ms | 0.0164 ms | 0.0154 ms |     	- |     - |     - |       13 B |
|  ArrayWithItems | 38.717 ms | 1.4343 ms | 1.2715 ms |         - |     - |     - |          - |
|     EmptyObject |  6.062 ms | 0.0517 ms | 0.0432 ms | 1140.6250 |     - |     - |  4800008 B |
| ObjectWithItems | 89.798 ms | 0.5728 ms | 0.5078 ms | 4833.3333 |     - |     - | 20800121 B |

Root Cache:

|          Method |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|---------------- |----------:|----------:|----------:|----------:|------:|------:|-----------:|
|            Null |  1.385 ms | 0.0118 ms | 0.0110 ms |         - |     - |     - |          - |
|           False |  1.415 ms | 0.0130 ms | 0.0108 ms |         - |     - |     - |        3 B |
|            True |  1.387 ms | 0.0116 ms | 0.0109 ms |         - |     - |     - |        9 B |
|          Number |  6.214 ms | 0.0496 ms | 0.0440 ms |         - |     - |     - |          - |
|          String |  7.238 ms | 0.0801 ms | 0.0749 ms |         - |     - |     - |          - |
|      EmptyArray |  2.487 ms | 0.0331 ms | 0.0310 ms |         - |     - |     - |        3 B |
|  ArrayWithItems | 42.533 ms | 0.3250 ms | 0.2881 ms |         - |     - |     - |          - |
|     EmptyObject |  4.629 ms | 0.0335 ms | 0.0297 ms | 1140.6250 |     - |     - |  4800010 B |
| ObjectWithItems | 92.924 ms | 1.1016 ms | 0.9765 ms | 4833.3333 |     - |     - | 20800121 B |

Removed Object Alloction For Keys:

|          Method |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |----------:|----------:|----------:|----------:|------:|------:|----------:|
|            Null |  1.379 ms | 0.0079 ms | 0.0070 ms |         - |     - |     - |         - |
|           False |  1.513 ms | 0.0108 ms | 0.0101 ms |         - |     - |     - |       9 B |
|            True |  1.392 ms | 0.0093 ms | 0.0082 ms |         - |     - |     - |         - |
|          Number |  6.390 ms | 0.1218 ms | 0.1354 ms |         - |     - |     - |         - |
|          String |  7.220 ms | 0.0779 ms | 0.0729 ms |         - |     - |     - |         - |
|      EmptyArray |  2.480 ms | 0.0210 ms | 0.0186 ms |         - |     - |     - |       3 B |
|  ArrayWithItems | 41.832 ms | 0.2922 ms | 0.2590 ms |         - |     - |     - |         - |
|     EmptyObject |  4.785 ms | 0.0407 ms | 0.0381 ms | 1140.6250 |     - |     - | 4800010 B |
| ObjectWithItems | 68.335 ms | 0.9880 ms | 0.9242 ms | 1125.0000 |     - |     - | 4800000 B |

Number64 To DoubleEx Direct:

|          Method |      Mean |     Error |    StdDev |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |----------:|----------:|----------:|----------:|------:|------:|----------:|
|            Null |  1.364 ms | 0.0219 ms | 0.0183 ms |         - |     - |     - |         - |
|           False |  1.411 ms | 0.0071 ms | 0.0059 ms |         - |     - |     - |       9 B |
|            True |  1.386 ms | 0.0106 ms | 0.0099 ms |         - |     - |     - |       9 B |
|          Number |  4.891 ms | 0.0290 ms | 0.0271 ms |         - |     - |     - |         - |
|          String |  7.109 ms | 0.1041 ms | 0.0869 ms |         - |     - |     - |         - |
|      EmptyArray |  2.497 ms | 0.0242 ms | 0.0226 ms |         - |     - |     - |       3 B |
|  ArrayWithItems | 39.660 ms | 0.1669 ms | 0.1479 ms |         - |     - |     - |         - |
|     EmptyObject |  4.618 ms | 0.0308 ms | 0.0273 ms | 1140.6250 |     - |     - | 4800010 B |
| ObjectWithItems | 67.946 ms | 0.6656 ms | 0.6226 ms | 1125.0000 |     - |     - | 4800258 B |
```

The conclusions:

* Null, False, True -> ~50% reduction
* Number -> ~33% reduction
* String -> ~14% reduction
* Empty Array -> 14% reduction
* Empty Object -> ~20% reduction
* Array -> ~ 6% slower
* Object -> ~25% reduction